+++ docs/src/markdown/about/changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.2.0
+
+- **NEW**: `--wait` removed from CLI as the server should not get hung while waiting. `wait` is still available when
+  using `pyluxa4.usb` directly.
+- **NEW**: Add new `schedule` command that takes a JSON and schedules lights by days and times.
+
 ## 1.1.0
 
 - **NEW**: Allow `traffic-light`, `police`, `rainbow`, and `random<1 - 5>` as valid values on CLI along with numerical